+++ packages/ui/src/components/NotificationHub/NotificationHub.tsx
@@ -2,9 +2,9 @@ import { XMarkIcon } from '@heroicons/react/24/solid';
 import { AnimatePresence, motion } from 'framer-motion';
 import { useTranslation } from 'react-i18next';
 
-import { useNotificationsStore } from '../../stores/notifications-store';
-import { withI18nProvider } from '../../utils/with-i18n-provider';
-import { Card } from '../Card/Card';
+import { useNotificationsStore } from '@/hooks/useNotificationsStore';
+import { Card } from '@/legacy/components/Card';
+
 import { NotificationIcon } from './NotificationIcon';
 
 type NotificationHubProps = {
@@ -12,7 +12,7 @@ type NotificationHubProps = {
   timeout?: number;
 };
 
-const NotificationHubComponent = ({ timeout = 5000 }: NotificationHubProps) => {
+const NotificationHub = ({ timeout = 5000 }: NotificationHubProps) => {
   const { t } = useTranslation();
   const { dismissNotification, notifications } = useNotificationsStore();
 
@@ -47,7 +47,7 @@ const NotificationHubComponent = ({ timeout = 5000 }: NotificationHubProps) => {
                       <XMarkIcon aria-hidden="true" className="h-5 w-5" />
                     </button>
                   </div>
-                  <p className="my-2 text-default">{item.message}</p>
+                  <p className="my-2 text-muted-foreground">{item.message}</p>
                 </div>
                 <motion.div
                   animate={{ width: '100%' }}
@@ -66,7 +66,5 @@ const NotificationHubComponent = ({ timeout = 5000 }: NotificationHubProps) => {
     </div>
   );
 };
-
-const NotificationHub = withI18nProvider(NotificationHubComponent) as typeof NotificationHubComponent;
 
 export { NotificationHub, type NotificationHubProps };