+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @douglasneuroinformatics/ui
 
+## 6.0.0
+
+### BREAKING CHANGES
+
+- Remove `transformTranslations`
+- Remove `LoadingFallback`, `FormPageWrapper`
+- Remove `text-strong` and `text-default` css classes
+- Import of `@douglasneuroinformatics/ui` has been removed. Instead, import from `@douglasneuroinformatics/ui/components`, `@douglasneuroinformatics/ui/hooks`, `@douglasneuroinformatics/ui/legacy`, `@douglasneuroinformatics/ui/i18n`, or `@douglasneuroinformatics/ui/utils`
+
 ## 5.2.0
 
 ### Minor Changes